# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "XSecurity-Vault",
    "description": "Secure digital certificate management orchestrator",
    "version": "2.0.0"
  },
  "paths": {
...
    "/regras-acesso/": {
      "get": {
        "tags": [
          "Regras de Acesso"
        ],
        "summary": "Listar",
        "operationId": "listar_regras_acesso__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RegraAcessoOut"
                  },
                  "type": "array",
                  "title": "Response Listar Regras Acesso  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Regras de Acesso"
        ],
        "summary": "Criar",
        "operationId": "criar_regras_acesso__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegraAcessoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegraAcessoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/regras-acesso/grupo/{grupo_id}": {
      "get": {
        "tags": [
          "Regras de Acesso"
        ],
        "summary": "Listar Grupo",
        "operationId": "listar_grupo_regras_acesso_grupo__grupo_id__get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegraAcessoOut"
                  },
                  "title": "Response Listar Grupo Regras Acesso Grupo  Grupo Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/regras-acesso/{regra_id}": {
      "get": {
        "tags": [
          "Regras de Acesso"
        ],
        "summary": "Obter",
        "operationId": "obter_regras_acesso__regra_id__get",
        "parameters": [
          {
            "name": "regra_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Regra Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegraAcessoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Regras de Acesso"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_regras_acesso__regra_id__put",
        "parameters": [
          {
            "name": "regra_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Regra Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegraAcessoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegraAcessoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Regras de Acesso"
        ],
        "summary": "Deletar",
        "operationId": "deletar_regras_acesso__regra_id__delete",
        "parameters": [
          {
            "name": "regra_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Regra Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/regras-acesso-hosts/": {
      "get": {
        "tags": [
          "Regras de Acesso (Hosts)"
        ],
        "summary": "Listar",
        "operationId": "listar_regras_acesso_hosts__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RegraAcessoHostOut"
                  },
                  "type": "array",
                  "title": "Response Listar Regras Acesso Hosts  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Regras de Acesso (Hosts)"
        ],
        "summary": "Criar",
        "operationId": "criar_regras_acesso_hosts__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegraAcessoHostCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegraAcessoHostOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/regras-acesso-hosts/grupo/{grupo_id}": {
      "get": {
        "tags": [
          "Regras de Acesso (Hosts)"
        ],
        "summary": "Listar Grupo",
        "operationId": "listar_grupo_regras_acesso_hosts_grupo__grupo_id__get",
        "parameters": [
          {
            "name": "grupo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grupo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegraAcessoHostOut"
                  },
                  "title": "Response Listar Grupo Regras Acesso Hosts Grupo  Grupo Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/regras-acesso-hosts/{regra_id}": {
      "get": {
        "tags": [
          "Regras de Acesso (Hosts)"
        ],
        "summary": "Obter",
        "operationId": "obter_regras_acesso_hosts__regra_id__get",
        "parameters": [
          {
            "name": "regra_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Regra Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegraAcessoHostOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Regras de Acesso (Hosts)"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_regras_acesso_hosts__regra_id__put",
        "parameters": [
          {
            "name": "regra_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Regra Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegraAcessoHostUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegraAcessoHostOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Regras de Acesso (Hosts)"
        ],
        "summary": "Deletar",
        "operationId": "deletar_regras_acesso_hosts__regra_id__delete",
        "parameters": [
          {
            "name": "regra_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Regra Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegraAcessoCreate": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "tipo_dia": {
            "$ref": "#/components/schemas/TipoDiaEnum"
          },
          "dias_especificos": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dias Especificos"
          },
          "horarios": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Horarios"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "tipo_dia",
          "horarios"
        ],
        "title": "RegraAcessoCreate",
        "description": "Modelo usado no POST /regras-acesso"
      },
      "RegraAcessoHostCreate": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "tipo_dia": {
            "$ref": "#/components/schemas/TipoDiaEnum"
          },
          "dias_especificos": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dias Especificos"
          },
          "horarios": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Horarios"
          },
          "urls": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urls"
          },
          "bloquear_em_feriado": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bloquear Em Feriado",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "tipo_dia",
          "horarios"
        ],
        "title": "RegraAcessoHostCreate"
      },
      "RegraAcessoHostOut": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "tipo_dia": {
            "$ref": "#/components/schemas/TipoDiaEnum"
          },
          "dias_especificos": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dias Especificos"
          },
          "horarios": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Horarios"
          },
          "urls": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urls"
          },
          "bloquear_em_feriado": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bloquear Em Feriado",
            "default": false
          },
          "regra_id": {
            "type": "integer",
            "title": "Regra Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "tipo_dia",
          "horarios",
          "regra_id",
          "criado_em"
        ],
        "title": "RegraAcessoHostOut"
      },
      "RegraAcessoHostUpdate": {
        "properties": {
          "tipo_dia": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TipoDiaEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "dias_especificos": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dias Especificos"
          },
          "horarios": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horarios"
          },
          "urls": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urls"
          },
          "bloquear_em_feriado": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bloquear Em Feriado"
          }
        },
        "type": "object",
        "title": "RegraAcessoHostUpdate"
      },
      "RegraAcessoOut": {
        "properties": {
          "grupo_id": {
            "type": "integer",
            "title": "Grupo Id"
          },
          "tipo_dia": {
            "$ref": "#/components/schemas/TipoDiaEnum"
          },
          "dias_especificos": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dias Especificos"
          },
          "horarios": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Horarios"
          },
          "regra_id": {
            "type": "integer",
            "title": "Regra Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          }
        },
        "type": "object",
        "required": [
          "grupo_id",
          "tipo_dia",
          "horarios",
          "regra_id",
          "criado_em"
        ],
        "title": "RegraAcessoOut"
      },
      "RegraAcessoUpdate": {
        "properties": {
          "tipo_dia": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TipoDiaEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "dias_especificos": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dias Especificos"
          },
          "horarios": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horarios"
          }
        },
        "type": "object",
        "title": "RegraAcessoUpdate"
      },
...
      "TipoDiaEnum": {
        "type": "string",
        "enum": [
          "corridos",
          "uteis",
          "especificos"
        ],
        "title": "TipoDiaEnum"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "Session token for authentication. Use the session_token from login response.",
        "scheme": "bearer"
      }
    }
  }
}